Code edited.

+++ src/Week2/Practices/Practice06_HarmonicMeanOfArray.java
@@ -4,8 +4,8 @@ public class Practice06_HarmonicMeanOfArray {
     public static void main(String[] args) {
         int[] numbers = {1, 2, 3, 4, 5};
         double harmonicMean = 0;
-        for (int i : numbers) {
-            harmonicMean += (double) 1/i;
+        for (double i : numbers) {
+            harmonicMean += 1/i;
         }
         System.out.println("Harmonik Ortalama: " + harmonicMean);
     }
